Create Stop war22 in Ukraine…!
While Ukraine is under missile attacks GitHub could be used by Russians to develop apps and platforms aiming to destabilize Ukrainian web resources.  
Please, prevent these actions and don't stay on the same side with invaders! All information about war can be found at:  https://war.ukraine.ua/   
We urge you to close GitHub for Russia and its developers! We value your support and we are in need for your actions!

It is not a shame, it is a disgust. The balance of peace in the world is ruined. Nobody knows the end of this cruelty. But humans. Real humans will not wash their hands in blood, will you? If Yes, this will be also blood of your kids, friends, parents. In future. 

Long live Ukraine.